python/requirements.txt
GitPython==0.1.7
boto==2.29.1
gitconfig==0.0.2
python-cloudfiles==1.7.11
wsgiref==0.1.2
